+++ scif/main/commands.py
@@ -1,21 +1,14 @@
-'''
+"""
 
-Copyright (C) 2017-2019 Vanessa Sochat.
+Copyright (C) 2017-2020 Vanessa Sochat.
 
-This program is free software: you can redistribute it and/or modify it
-under the terms of the GNU Affero General Public License as published by
-the Free Software Foundation, either version 3 of the License, or (at your
-option) any later version.
+This Source Code Form is subject to the terms of the
+Mozilla Public License, v. 2.0. If a copy of the MPL was not distributed
+with this file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-This program is distributed in the hope that it will be useful, but WITHOUT
-ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
-FITNESS FOR A PARTICULAR PURPOSE.  See the GNU Affero General Public
-License for more details.
+Modified from https://github.com/Visual-mov/Colorful-Julia (MIT License)
 
-You should have received a copy of the GNU Affero General Public License
-along with this program.  If not, see <https://www.gnu.org/licenses/>.
-
-'''
+"""
 
 from scif.logger import bot
 from scif.defaults import SCIF_BASE
@@ -26,7 +19,7 @@ import os
 
 
 def _exec(self, app=None, interactive=False, exit=False):
-    '''exec is the underlying driver of both run and exec, taking a final
+    """exec is the underlying driver of both run and exec, taking a final
        SCIF and executing the command for the user.
 
        This function is called via self._exec, and the expectation is that
@@ -42,11 +35,11 @@ def _exec(self, app=None, interactive=False, exit=False):
        interactive: if True, us os.system directly
        exit: exit with return code from command (for test)
 
-    '''
+    """
 
-    name = ''
+    name = ""
     if app is not None and app in self.apps():
-        name = '[%s] ' %app
+        name = "[%s] " % app
 
     # If the entry folder still not set, we don't have an app
     if self._entry_folder is None:
@@ -56,30 +49,30 @@ def _exec(self, app=None, interactive=False, exit=False):
     os.chdir(self._entry_folder)
 
     # export environment
-    runtime_environ = self.export_env() 
+    runtime_environ = self.export_env()
 
     # Execv to the entrypoint
     executable = which(self._entry_point[0])
 
     # Exit if command not found
-    args = ''
+    args = ""
     if executable is None:
         bot.exit("%s not found." % self._entry_point[0])
 
     if len(self._entry_point) > 1:
         if exit is False:
-            args = ' '.join(self._entry_point[1:])
+            args = " ".join(self._entry_point[1:])
         else:
             args = self._entry_point[1:]
 
     # Are we executing a command as a string?
     if not isinstance(args, list):
-        cmd = "%s %s" %(executable, args)
-        bot.info('%sexecuting %s' %(name, cmd))
+        cmd = "%s %s" % (executable, args)
+        bot.info("%sexecuting %s" % (name, cmd))
 
     # or a list with execv?
     else:
-        bot.info('%sexecuting %s %s' %(name, executable, ' '.join(args)))
+        bot.info("%sexecuting %s %s" % (name, executable, " ".join(args)))
 
     # Return output to console
     loc = locale.getdefaultlocale()[1]
@@ -89,12 +82,12 @@ def _exec(self, app=None, interactive=False, exit=False):
 
         # Will exit with error, if happens, otherwise exit with 0
         if exit is True:
-            result = self._run_command(cmd=[executable] + args, 
-                                       spinner=False, 
-                                       quiet=self.quiet)
-            sys.exit(result['return_code'])
+            result = self._run_command(
+                cmd=[executable] + args, spinner=False, quiet=self.quiet
+            )
+            sys.exit(result["return_code"])
         else:
-            os.system(''.join(cmd))
+            os.system("".join(cmd))
 
     else:
         for line in os.popen(cmd):
@@ -105,7 +98,7 @@ def _exec(self, app=None, interactive=False, exit=False):
 
 
 def execute(self, app, cmd=None, args=None):
-    '''execute a command in the context of an app. This means the following:
+    """execute a command in the context of an app. This means the following:
 
     1. Check that the app is valid for the client. Don't proceed otherwise
     2. Set the client app to be active
@@ -116,19 +109,19 @@ def execute(self, app, cmd=None, args=None):
     ==========
     app: the name of the scif app to execute a command to
 
-    '''
+    """
     self.activate(app, cmd, args)
-                          # checks for existence
-                          # sets _active to app's name
-                          # updates environment
-                          # sets entrypoint
-                          # sets entryfolder
+    # checks for existence
+    # sets _active to app's name
+    # updates environment
+    # sets entrypoint
+    # sets entryfolder
 
     return self._exec(app)
 
 
 def shell(self, app, cmd=None):
-    '''akin to execute, but specific for shell. In this case, we pass the
+    """akin to execute, but specific for shell. In this case, we pass the
        calling process to the shell. We do the same steps as in run/exec:
 
     1. Check that the app is valid for the client. Don't proceed otherwise
@@ -140,25 +133,24 @@ def shell(self, app, cmd=None):
     ==========
     app: the name of the scif app to execute a command to
 
-    '''
+    """
     from scif.defaults import SCIF_SHELL
 
     if cmd is None:
         cmd = SCIF_SHELL
 
     self.activate(app, cmd)
-                            # checks for existence
-                            # sets _active to app's name
-                            # updates environment
-                            # sets shell entrypoint
+    # checks for existence
+    # sets _active to app's name
+    # updates environment
+    # sets shell entrypoint
 
-                            # interactive
+    # interactive
     return self._exec(app, True)
 
 
-
 def run(self, app=None, args=None):
-    '''run an app. This means the following:
+    """run an app. This means the following:
 
     1. Check that the app is valid for the client. Don't proceed otherwise
     2. Set the client app to be active
@@ -170,24 +162,24 @@ def run(self, app=None, args=None):
     app: the name of the scif app to run
     args: a list of one or more additional arguments to pass to runscript
 
-    '''
+    """
     interactive = False
     config = self.app(app)
-    if 'apprun' not in config:
-        bot.debug('%s does not have a runscript.' % app)
+    if "apprun" not in config:
+        bot.debug("%s does not have a runscript." % app)
         interactive = True
 
-    self.activate(app, args=args)    # checks for existence
-                                     # sets _active to app's name
-                                     # updates environment
-                                     # sets entrypoint
-                                     # sets entryfolder
+    self.activate(app, args=args)  # checks for existence
+    # sets _active to app's name
+    # updates environment
+    # sets entrypoint
+    # sets entryfolder
 
     return self._exec(app, interactive=interactive)
 
 
 def test(self, app=None, args=None):
-    '''test an app. This means the following:
+    """test an app. This means the following:
 
     1. Check that the app is valid for the client. Don't proceed otherwise
     2. Set the client app to be active
@@ -200,7 +192,7 @@ def test(self, app=None, args=None):
     app: the name of the scif app to run. If none provided, all apps are tested.
     args: a list of one or more additional arguments to pass to runscript
 
-    '''
+    """
 
     # Does the application have a test script?
 
@@ -208,7 +200,7 @@ def test(self, app=None, args=None):
 
         self.activate(app, args=args)
 
-        if not self._set_entrypoint(app, 'SCIF_APPTEST', args):
+        if not self._set_entrypoint(app, "SCIF_APPTEST", args):
             bot.info("No tests defined for this app.")
             sys.exit(1)